changed Component name and added defaults
