+++ src/index.js
@@ -51,7 +51,7 @@ export { default as ToggleColorMode } from "./ToggleColorMode";
 export { default as TransferList } from "./TransferList";
 export { default as UserMenu } from "./UserMenu";
 export { UserAvatar, useColorMap } from "./UserAvatar";
-export { default as UseConfirmDialog } from "./UseConfirmDialog";
+export { default as ConfirmProvider, useConfirm } from "./ConfirmProvider";
 export {
   IpgLogo,
   VirtoLogo,
